refactor: add MealIcon and NoFoodText components; update DailyGoal and MealFormAccordionItem for localization
