+++ src/components/NoFoodText/NoFoodText.module.css
@@ -1,0 +1,6 @@
+.text {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
